mytube/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-cors-headers==3.13.0
django-crontab==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
idna==3.4
pycodestyle==2.9.1
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.12
